+++ examples/webdriverio-mocha-todomvc/spec/manage_a_todo_list.spec.ts
@@ -10,10 +10,6 @@ describe('Managing a Todo List', () => {
             ClearLocalStorage(),
         ));
 
-    // afterEach(() => {
-    //     browser.execute(`window.localStorage.clear()`);
-    // });
-
     describe('TodoMVC', () => {
 
         describe('actor', () => {