+++ packages/jasmine/src/adapter/JasmineAdapter.ts
@@ -80,7 +80,17 @@ export class JasmineAdapter implements TestRunnerAdapter {
         ));
 
         this.runner.addReporter(reporter(this.runner.jasmine));
+
+        if (this.config.reporters) {
+            this.config.reporters.forEach(reporter => {
+                this.runner.addReporter(reporter)
+            });
+        }
+
         this.runner.configureDefaultReporter(this.config);
+
+        this.runner.loadRequires();
+        this.runner.loadHelpers();
 
         this.configureSpecFilter();