+++ CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.31.1](https://github.com/serenity-js/serenity-js/compare/v2.31.0...v2.31.1) (2021-08-27)
+
+
+### Bug Fixes
+
+* **cucumber:** don't pass the "rerun" file to Cucumber if it doesn't exist ([b08eca2](https://github.com/serenity-js/serenity-js/commit/b08eca2b849194835385d0966b0f4a9895fe1d24)), closes [#971](https://github.com/serenity-js/serenity-js/issues/971) [protractor-cucumber-framework/protractor-cucumber-framework#219](https://github.com/protractor-cucumber-framework/protractor-cucumber-framework/issues/219)
+* **cucumber:** fixed duplicate "rerun" param passed to Cucumber ([8d93788](https://github.com/serenity-js/serenity-js/commit/8d937885ce1f230c41cb3b40c0e7026c859cbad6)), closes [#971](https://github.com/serenity-js/serenity-js/issues/971)
+
+
+
+
+
+# [2.31.0](https://github.com/serenity-js/serenity-js/compare/v2.30.3...v2.31.0) (2021-08-23)
+
+
+### Bug Fixes
+
+* **jasmine:** added JasmineReporter interface ([c72d03b](https://github.com/serenity-js/serenity-js/commit/c72d03bbb4976ccb980144cf99f4ea8dbc5193bc)), closes [#962](https://github.com/serenity-js/serenity-js/issues/962)
+
+
+### Features
+
+* **jasmine:** support additional reporters injected in confg (i.e. ReportPortal) ([e2b102a](https://github.com/serenity-js/serenity-js/commit/e2b102a573878596afb7f8b7424843b468961b97))
+
+
+
+
+
+## [2.30.3](https://github.com/serenity-js/serenity-js/compare/v2.30.2...v2.30.3) (2021-08-16)
+
+
+### Bug Fixes
+
+* **jasmine:** ensure JasmineAdapter loads configured "requires" and "helpers" ([5dfd45c](https://github.com/serenity-js/serenity-js/commit/5dfd45cd5edcc36fe98a07d5abf1f2abfe9f2624)), closes [#954](https://github.com/serenity-js/serenity-js/issues/954)
+
+
+
+
+
+## [2.30.2](https://github.com/serenity-js/serenity-js/compare/v2.30.1...v2.30.2) (2021-08-13)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @babel/parser to ^7.15.3 ([dbed99e](https://github.com/serenity-js/serenity-js/commit/dbed99e4004aac0719faf0b3369d9977383de87c))
+* **deps:** update dependency find-java-home to ^1.2.2 ([58cb156](https://github.com/serenity-js/serenity-js/commit/58cb1566517d055a9ba885eaafed29b07acda08f))
+* **mocha:** upgraded Mocha to version 9 ([446c7b9](https://github.com/serenity-js/serenity-js/commit/446c7b9f3440303a3bfaf4197ca3cb1aeb1a72e5))
+
+
+
+
+
 ## [2.30.1](https://github.com/serenity-js/serenity-js/compare/v2.30.0...v2.30.1) (2021-08-09)