+++ packages/jasmine/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.31.1](https://github.com/serenity-js/serenity-js/compare/v2.31.0...v2.31.1) (2021-08-27)
+
+**Note:** Version bump only for package @serenity-js/jasmine
+
+
+
+
+
+# [2.31.0](https://github.com/serenity-js/serenity-js/compare/v2.30.3...v2.31.0) (2021-08-23)
+
+
+### Bug Fixes
+
+* **jasmine:** added JasmineReporter interface ([c72d03b](https://github.com/serenity-js/serenity-js/commit/c72d03bbb4976ccb980144cf99f4ea8dbc5193bc)), closes [#962](https://github.com/serenity-js/serenity-js/issues/962)
+
+
+### Features
+
+* **jasmine:** support additional reporters injected in confg (i.e. ReportPortal) ([e2b102a](https://github.com/serenity-js/serenity-js/commit/e2b102a573878596afb7f8b7424843b468961b97))
+
+
+
+
+
+## [2.30.3](https://github.com/serenity-js/serenity-js/compare/v2.30.2...v2.30.3) (2021-08-16)
+
+
+### Bug Fixes
+
+* **jasmine:** ensure JasmineAdapter loads configured "requires" and "helpers" ([5dfd45c](https://github.com/serenity-js/serenity-js/commit/5dfd45cd5edcc36fe98a07d5abf1f2abfe9f2624)), closes [#954](https://github.com/serenity-js/serenity-js/issues/954)
+
+
+
+
+
+## [2.30.2](https://github.com/serenity-js/serenity-js/compare/v2.30.1...v2.30.2) (2021-08-13)
+
+**Note:** Version bump only for package @serenity-js/jasmine
+
+
+
+
+
 ## [2.30.1](https://github.com/serenity-js/serenity-js/compare/v2.30.0...v2.30.1) (2021-08-09)
 
 **Note:** Version bump only for package @serenity-js/jasmine